+++ code/src/lib/processing/Processor.cpp
@@ -305,9 +305,9 @@ void Processor::processPointCloud(const PointCloud<double, 3>& pointCloud) {
         Helpers::buildFactorGraph(mDEM, mGraph, mCoefficients, mVariances,
           mWeights, factorGraph, fgMapping);
         PropertySet opts;
-        opts.set("maxiter", (size_t)mMaxBPIter);
+        opts.set("maxiter", mMaxBPIter);
         opts.set("tol", mBPTol);
-        opts.set("verbose", (size_t)0);
+        opts.set("verbose", 0);
         opts.set("updates", std::string("SEQRND"));
         opts.set("logdomain", mLogDomain);
         opts.set("inference", std::string("MAXPROD"));
